dr: Load text synchronously
This seems to be more SEO-friendly.

I also experimented with using CSS variables to selectively display
translated text instead of re-rendering everything but this turned
out to be *way* slower, twenty seconds to process a language
change. (I'm using this trick for echoflower but I guess I'll have to
look for something else when I scale that one up to the full text.)

Hopefully adding the "(JP)" link back will allow the Japanese text to
be indexed as well.