dr: Make links lazy to reduce memory use
Creating the little floating link next to each line only when it's
actually hovered over reduces the page's memory use by about 40%.

The page is still heavy after this optimization. Most of the memory
use is from the sheer size of the DOM tree. To make the page really
snappy I'll probably have to move to full lazy rendering and custom
page search...