dr: Fix stray letter "j", make / and % handling more principled
An anonymous bug reporter pointed out that
obj_dw_churchb_rotatingtower_slash_Create_0_gml_90_0 rendered as

> * Written somewhere no one can erase!j

This is because the message ends with `/%j`. The "j" isn't rendered
because `/` ends the display. Probably a typo. (It's nevertheless
carried over to the Japanese translation.)

Now we stop rendering after a `/` or `%` and assert that there's no
text after it. This turned up a few new entries for the `%` allowlist,
i.e. messages that should be rendered with a percent sign but weren't.